[refactor] Split off kmsauth client to new file
